start moving glview3d implementation into viewerwindow
