+++ renderlib/CMakeLists.txt
@@ -17,6 +17,7 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppScene.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppScene.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/CCamera.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/CCamera.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/command.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/command.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/Flags.cpp"
@@ -47,6 +48,8 @@ target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/tiny_obj_loader.cc"
 	"${CMAKE_CURRENT_SOURCE_DIR}/tiny_obj_loader.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/version.hpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/ViewerWindow.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/ViewerWindow.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/VolumeDimensions.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/VolumeDimensions.h"
 )
