+++ renderlib/Manipulator.cpp
@@ -1,10 +1,1 @@
 #include "Manipulator.h"
-
-void
-RotateManipulator::drawGL()
-{
-}
-void
-RotateManipulator::pickGL(int x, int y)
-{
-}
